requirements.txt
python-telegram-bot==21.3
aiohttp==3.9.1
python-dotenv==1.0.0
Pillow==10.3.0
websockets==12.0
